requirements.txt
asgiref==3.5.2
boto3==1.24.15
botocore==1.27.15
Django==4.0.5
django-ckeditor==6.4.2
django-js-asset==2.0.0
django-quill==0.1.0
django-storages==1.12.3
django-tinymce==3.4.0
jmespath==1.0.1
Pillow==9.1.1
psycopg2-binary==2.9.3
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.2.0
